abstracted out engine for chunkypng
