+++ lib/compass/sass_extensions/sprites/base.rb
@@ -1,121 +1,156 @@
 module Compass
   module SassExtensions
     module Sprites
-      class Base < Sass::Script::Literal
-        # Changing this string will invalidate all previously generated sprite images.
-        # We should do so only when the packing algorithm changes
-        SPRITE_VERSION = "1"
-
-        attr_accessor :image_names, :path, :name, :options
-        attr_accessor :images, :width, :height
-
-        def self.from_uri(uri, context, kwargs)
-          path, name = Compass::Sprites.path_and_name(uri.value)
-          sprites = Compass::Sprites.discover_sprites(uri.value).map do |sprite|
-            sprite.gsub(Compass.configuration.images_path+"/", "")
+      module Base
+        module ClassMethods
+          def from_uri(uri, context, kwargs)
+            path, name = Compass::Sprites.path_and_name(uri.value)
+            sprites = Compass::Sprites.discover_sprites(uri.value).map do |sprite|
+              sprite.gsub(Compass.configuration.images_path+"/", "")
+            end
+            new(sprites, path, name, context, kwargs)
           end
-          new(sprites, path, name, context, kwargs)
         end
+        #instance Methods
+        module InstanceMethods
+          
+          def self.included(base)
+            base.extend(ClassMethods)
+          end
+          
+          # Changing this string will invalidate all previously generated sprite images.
+          # We should do so only when the packing algorithm changes
+          SPRITE_VERSION = "1"
 
-        def initialize(image_names, path, name, context, options)
-          @image_names, @path, @name, @options = image_names, path, name, options
-          @images = nil
-          @width = nil
-          @height = nil
-          @evaluation_context = context
-          validate!
-          compute_image_metadata!
-        end
+          attr_accessor :image_names, :path, :name, :options
+          attr_accessor :images, :width, :height
 
-        # Calculate the size of the sprite
-        def size
-          [width, height]
-        end
 
-        def sprite_names
-          image_names.map{|f| Compass::Sprites.sprite_name(f) }
-        end
+          def initialize(image_names, path, name, context, options)
+            @image_names, @path, @name, @options = image_names, path, name, options
+            @images = nil
+            @width = nil
+            @height = nil
+            @evaluation_context = context
+            validate!
+            compute_image_metadata!
+          end
 
-        def validate!
-          for sprite_name in sprite_names
-            unless sprite_name =~ /\A#{Sass::SCSS::RX::IDENT}\Z/
-              raise Sass::SyntaxError, "#{sprite_name} must be a legal css identifier"
+          # Calculate the size of the sprite
+          def size
+            [width, height]
+          end
+        
+          # Calculates the overal image dimensions
+          # collects image sizes and input parameters for each sprite
+          def compute_image_metadata!
+            @width = 0
+            init_images
+            compute_image_positions!
+            @height = @images.last.top + @images.last.height
+          end
+
+          def init_images
+            @images = image_names.collect do |relative_file|
+              image = Compass::SassExtensions::Sprites::Image.new(relative_file, options)
+              @width = [ @width, image.width + image.offset ].max
+              image
             end
           end
-        end
-
-        # The on-the-disk filename of the sprite
-        def filename
-          File.join(Compass.configuration.images_path, "#{path}-#{uniqueness_hash}.png")
-        end
-
-        # Calculates the overal image dimensions
-        # collects image sizes and input parameters for each sprite
-        def compute_image_metadata!
-        end
-
-        # Generate a sprite image if necessary
-        def generate
-          if generation_required?
-            sprite_data = construct_sprite
-            save!(sprite_data)
-            Compass.configuration.run_callback(:sprite_generated, sprite_data)
+          # Calculates the overal image dimensions
+          # collects image sizes and input parameters for each sprite
+          def compute_image_positions!
+            @images.each_with_index do |image, index|
+              image.left = image.position.unit_str == "%" ? (@width - image.width) * (image.position.value / 100) : image.position.value
+              next if index == 0
+              last_image = @images[index-1]
+              image.top = last_image.top + last_image.height + [image.spacing,  last_image.spacing].max
+            end
           end
-        end
 
-        def generation_required?
-          !File.exists?(filename) || outdated?
-        end
+          def image_for(name)
+            @images.detect { |img| img.name == name}
+          end
 
-        def uniqueness_hash
-          @uniqueness_hash ||= begin
-            sum = Digest::MD5.new
-            sum << SPRITE_VERSION
-            sum << path
-            images.each do |image|
-              [:relative_file, :height, :width, :repeat, :spacing, :position, :digest].each do |attr|
-                sum << image.send(attr).to_s
+          def sprite_names
+            image_names.map{|f| Compass::Sprites.sprite_name(f) }
+          end
+
+          def validate!
+            for sprite_name in sprite_names
+              unless sprite_name =~ /\A#{Sass::SCSS::RX::IDENT}\Z/
+                raise Sass::SyntaxError, "#{sprite_name} must be a legal css identifier"
               end
             end
-            sum.hexdigest[0...10]
           end
-          @uniqueness_hash
-        end
 
-        # saves the sprite for later retrieval
-        def save!(output_png)
-          saved = output_png.save filename
-          Compass.configuration.run_callback(:sprite_saved, filename)
-          saved
-        end
+          # The on-the-disk filename of the sprite
+          def filename
+            File.join(Compass.configuration.images_path, "#{path}-#{uniqueness_hash}.png")
+          end
 
-        # All the full-path filenames involved in this sprite
-        def image_filenames
-          image_names.map do |image_name|
-            File.join(Compass.configuration.images_path, image_name)
+          # Generate a sprite image if necessary
+          def generate
+            if generation_required?
+              sprite_data = construct_sprite
+              save!(sprite_data)
+              Compass.configuration.run_callback(:sprite_generated, sprite_data)
+            end
+          end
+
+          def generation_required?
+            !File.exists?(filename) || outdated?
+          end
+
+          def uniqueness_hash
+            @uniqueness_hash ||= begin
+              sum = Digest::MD5.new
+              sum << SPRITE_VERSION
+              sum << path
+              images.each do |image|
+                [:relative_file, :height, :width, :repeat, :spacing, :position, :digest].each do |attr|
+                  sum << image.send(attr).to_s
+                end
+              end
+              sum.hexdigest[0...10]
+            end
+            @uniqueness_hash
+          end
+
+          # saves the sprite for later retrieval
+          def save!(output_png)
+            saved = output_png.save filename
+            Compass.configuration.run_callback(:sprite_saved, filename)
+            saved
+          end
+
+          # All the full-path filenames involved in this sprite
+          def image_filenames
+            image_names.map do |image_name|
+              File.join(Compass.configuration.images_path, image_name)
+            end
+          end
+
+          def inspect
+            to_s
+          end
+
+          def to_s(options = self.options)
+            sprite_url(self).value
+          end
+
+          def respond_to?(meth)
+            super || @evaluation_context.respond_to?(meth)
+          end
+
+          def method_missing(meth, *args, &block)
+            if @evaluation_context.respond_to?(meth)
+              @evaluation_context.send(meth, *args, &block)
+            else
+              super
+            end
           end
         end
-
-        def inspect
-          to_s
-        end
-
-        def to_s(options = self.options)
-          sprite_url(self).value
-        end
-
-        def respond_to?(meth)
-          super || @evaluation_context.respond_to?(meth)
-        end
-
-        def method_missing(meth, *args, &block)
-          if @evaluation_context.respond_to?(meth)
-            @evaluation_context.send(meth, *args, &block)
-          else
-            super
-          end
-        end
-
       end
     end
   end
